Add optional fetchMoreOverride method to PaginationBase
